backend/requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
djangorestframework==3.15.1
Markdown==3.6
django-filter==24.2
python-dotenv==1.0.1
tzdata==2024.1
argon2-cffi==23.1.0
pyjwt==2.8.0
cryptography==42.0.7
channels[daphne]==4.1.0
channels-redis==4.2.0
requests==2.32.3
django-cors-headers==4.6.0
